Fixed Bug
Bug:
"Defective" Writing to File of The Worker's Wage.

Reason:
Conversion From "double" To "int" of The Wage's Value.

+++ Ugh_Eats_FINAL/People.cpp
@@ -319,7 +319,7 @@ unordered_set<Admin*, hashAdmin, eqAdmin> Admin::readAdmins(ifstream & workers_s
 		a.set_birthday(d);
 
 		getline(workers_stream, textline);
-		a.set_wage(stoi(textline));
+		a.set_wage(stod(textline));
 
 		getline(workers_stream, textline);
 		a.set_role(textline);
